Massive optimization by demoting some uint16_t to uint8_t in ir.cpp!
Saves 60 bytes of code and runtime. Why did the compiler do such a bad job loading them?  Look at this...

```
            if (sendpulse_spaces==0) {
     dfe:	80 91 55 01 	lds	r24, 0x0155	; 0x800155 <_ZL16sendpulse_spaces>
     e02:	90 91 56 01 	lds	r25, 0x0156	; 0x800156 <_ZL16sendpulse_spaces+0x1>
     e06:	89 2b       	or	r24, r25
     e08:	a1 f5       	brne	.+104    	; 0xe72 <__vector_10+0xaa>

           if (sendpulse_bitmask&0x01) SP_PIN_A_SET_1();
     e0a:	80 91 57 01 	lds	r24, 0x0157	; 0x800157 <_ZL17sendpulse_bitmask>
     e0e:	90 91 58 01 	lds	r25, 0x0158	; 0x800158 <_ZL17sendpulse_bitmask+0x1>
     e12:	80 fd       	sbrc	r24, 0
     e14:	72 9a       	sbi	0x0e, 2	; 14

```

Wow! Look at all those impotent loads into r25! How is this posisble that the compiler can do this?

Yikes! Stay away from uint16_t when ever possible!

+++ cores/blinkcore/ir.cpp
@@ -310,9 +310,9 @@ uint8_t ir_test_and_charge_cli( void ) {
     
 */
 
-static volatile uint16_t sendpulse_bitmask;      // Which IR LEDs to send on 
-static volatile uint16_t sendpulse_spaces;       // Time to delay until next pulse. 0=pulse sent
-static volatile uint16_t sendpulse_spaces_next;  // A one entry deep buffer for spaces so we can try to keep it primed. 
+static volatile uint8_t sendpulse_bitmask;      // Which IR LEDs to send on 
+static volatile uint8_t sendpulse_spaces;       // Time to delay until next pulse. 0=pulse sent
+static volatile uint8_t sendpulse_spaces_next;  // A one entry deep buffer for spaces so we can try to keep it primed. 
 
 // Currently clocks at 23us @ 4Mhz
